+++ tests/testthat/test-get_fia_tables.R
@@ -5,7 +5,7 @@ test_that("fia_download() works", {
   expect_true(any(stringr::str_detect(files, "RI_CSV.zip")))
   expect_true(any(stringr::str_detect(files, "RI_PLOTGEOM.csv")))
 
-  db <- read_fia(states = "RI", dir = path)
+  db <- fia_load(states = "RI", dir = path)
 
   expect_type(db, "list")
   expect_s3_class(db[[1]], "data.frame")